+++ app/serializers/orcid_account_serializer.rb
@@ -6,11 +6,21 @@ class OrcidAccountSerializer < ActiveModel::Serializer
     :status,
     :oauth_authorize_url
 
-  def id
-    # Returning nil for the ID is an effective way to disable the feature on the
-    # frontend. The connect to orcid button will remain disabled without an
-    # OrcidAccount.
-    return nil unless TahiEnv.orcid_connect_enabled?
-    object.id
+  private
+
+  def include_name?
+    is_current_user?
+  end
+
+  def include_oauth_authorize_url?
+    is_current_user?
+  end
+
+  def include_status?
+    is_current_user?
+  end
+
+  def is_current_user?
+    current_user.id == object.user_id
   end
 end