+++ client/app/models/invitation.js
@@ -7,21 +7,30 @@ let currentState = function(stateName) {
 
 export default DS.Model.extend({
   abstract: DS.attr('string'),
-  attachments: DS.hasMany('invitation-attachment', { async: true }),
+  attachments: DS.hasMany('invitation-attachment'),
   body: DS.attr('string'),
   createdAt: DS.attr('date'),
   declineReason: DS.attr('string'),
   email: DS.attr('string'),
   information: DS.attr('string'),
   invitationType: DS.attr('string'),
-  invitee: DS.belongsTo('user', { inverse: 'invitations', async: true }),
+  invitee: DS.belongsTo('user', { inverse: 'invitations' }),
   inviteeRole: DS.attr('string'),
+  position: DS.attr('number'),
   primary: DS.belongsTo('invitation', { inverse: 'alternates', async: false }),
   alternates: DS.hasMany('invitation'),
   reviewerSuggestions: DS.attr('string'),
   state: DS.attr('string'),
-  task: DS.belongsTo('task', { polymorphic: true, async: true }),
+  task: DS.belongsTo('task', { polymorphic: true }),
+  decision: DS.belongsTo('decision', {async: false}),
   title: DS.attr('string'),
+
+  isAlternate: Ember.computed.notEmpty('primary'),
+  isPrimary: Ember.computed.notEmpty('alternates'),
+
+  canReposition: Ember.computed.notEmpty('validNewPositionsForInvitation'),
+
+  paperType: DS.attr('string'),
   updatedAt: DS.attr('date'),
   invitedAt: DS.attr('date'),
   acceptedAt: DS.attr('date'),
@@ -38,6 +47,8 @@ export default DS.Model.extend({
   declined: currentState('declined'),
   rescinded: currentState('rescinded'),
 
+  validNewPositionsForInvitation: DS.attr(), // will come in as an array of integers
+
   invitationFeedbackIsBlank: Ember.computed(
     'reviewerSuggestions',
     'declineReason',
@@ -46,6 +57,8 @@ export default DS.Model.extend({
         Ember.isBlank(this.get('declineReason'));
     }
   ),
+
+  isInvitedOrAccepted: Ember.computed.or('invited', 'accepted'),
 
   needsUserUpdate: Ember.computed.or('invited', 'pendingFeedback'),
 
@@ -60,6 +73,14 @@ export default DS.Model.extend({
       this.store.pushPayload(data);
       return this;
     });
+  },
+
+  updatePrimary(primaryId) {
+    return this.get('restless').putUpdate(this, '/update_primary', { primary_id: primaryId });
+  },
+
+  changePosition(newPosition) {
+    return this.get('restless').putUpdate(this, '/update_position', { position: newPosition });
   },
 
   decline() {
@@ -78,7 +99,7 @@ export default DS.Model.extend({
      });
   },
 
-  send() {
+  invite() {
     return this.get('restless').putUpdate(this, '/send_invite');
   },