+++ config/routes.rb
@@ -93,6 +93,8 @@ Tahi::Application.routes.draw do
       put :decline, on: :member
       put :rescind, on: :member
       put :send_invite, on: :member
+      put :update_position, on: :member
+      put :update_primary, on: :member
       get :details, on: :member
       resources :attachments, only: [:index, :create, :update, :destroy, :show], controller: 'invitation_attachments' do
         put :update_attachment, on: :member