+++ engines/tahi_standard_tasks/client/test-support/integration/components/invitation-detail-row-test.js
@@ -1,18 +1,17 @@
-import {moduleForComponent, test} from 'ember-qunit';
+import { moduleForComponent, test } from 'ember-qunit';
 import hbs from 'htmlbars-inline-precompile';
-import Ember from 'ember';
+import { manualSetup, make } from 'ember-data-factory-guy';
 
 moduleForComponent('invitation-detail-row', 'Integration | Component | invitation detail row', {
   integration: true,
   beforeEach() {
+    manualSetup(this.container);
     this.set('update-date', new Date('January 01, 2016'));
-    this.set('invitation', Ember.Object.create({
+    this.set('invitation', make('invitation', {
       declineReason: null,
       declined: false,
       email: 'jane@example.com',
-      invitee: Ember.Object.create({
-        fullName: 'Jane McEdits'
-      }),
+      invitee: { fullName: 'Jane McEdits' },
       reviewerSuggestions: null,
       state: 'pending',
       title: 'Awesome Paper!',
@@ -25,9 +24,8 @@ let template = hbs`{{invitation-detail-row
                       invitation=invitation
                       uiState='closed'}}`;
 
-test('displays invitation information if the invite.invited is true', function(assert){
-  this.set('invitation.pending', false);
-  this.set('invitation.invited', true);
+test('displays invitation information if invitation.invited is true', function(assert){
+  this.set('invitation.state', 'invited');
   this.set('invitation.invitedAt', this.get('update-date'));
   this.render(template);
 
@@ -50,50 +48,164 @@ test('displays invitation email when no invitee present', function(assert){
   assert.textPresent('.invitation-item-full-name', 'jane@example.com');
 });
 
-test('displays delete icon if invite is pending, the row is in the show state, and in current round', function(assert) {
+
+// header buttons
+
+let openTemplate = hbs`{{invitation-detail-row invitation=invitation
+                                               currentRound=currentRound
+                                               uiState=uiState}}`;
+
+test('the row is in the closed state, in the current round', function(assert) {
   this.setProperties({
-    'invitation.pending': true,
+    'invitation.state': 'pending',
+    currentRound: true,
+    uiState: 'closed'
+  });
+  this.render(openTemplate);
+  assert.elementNotFound('.invitation-item-action-delete', 'destroy button not shown if the row is closed');
+  assert.elementNotFound('.invitation-item-action-rescind', 'rescind not shown when the invitation is pending');
+  assert.elementFound('.invitation-item-action-send', 'send button shown when the invitation is pending');
+});
+
+test('the row is in the closed state, invitation is invited, in the current round', function(assert) {
+  this.setProperties({
+    'invitation.state': 'invited',
+    currentRound: true,
+    uiState: 'closed'
+  });
+  this.render(openTemplate);
+  assert.elementNotFound('.invitation-item-action-delete', 'destroy button not shown if the row is closed');
+  assert.elementNotFound('.invitation-item-action-rescind', 'rescind not shown when the invitation is invited');
+  assert.elementNotFound('.invitation-item-action-send', 'send button not shown when the invitation is invited');
+});
+
+test('the row is in the closed state, in the previous round', function(assert) {
+  this.setProperties({
+    'invitation.state': 'pending',
+    currentRound: false,
+    uiState: 'closed'
+  });
+  this.render(openTemplate);
+  assert.elementNotFound('.invitation-item-action-delete');
+  assert.elementNotFound('.invitation-item-action-rescind');
+  assert.elementNotFound('.invitation-item-action-send', 'cant send invites from previous rounds');
+
+});
+
+test('the row is in the show state, invitation is pending, and in current round', function(assert) {
+  this.setProperties({
+    'invitation.state': 'pending',
     currentRound: true,
     uiState: 'show'
   });
 
-  const openTemplate = hbs`{{invitation-detail-row allowDestroy=allowDestroy
-                                                   invitation=invitation
-                                                   currentRound=currentRound
-                                                   uiState=uiState}}`;
+  this.render(openTemplate);
+
+  assert.elementFound('.invitation-item-action-delete');
+  assert.elementNotFound('.invitation-item-action-rescind', 'rescind not shown when the invitation is pending');
+  assert.elementFound('.invitation-item-action-send', 'send button shown when the invitation is pending');
+});
+
+test('the row is in the show state, invitation is invited, and in current round', function(assert) {
+  this.setProperties({
+    'invitation.state': 'invited',
+    currentRound: true,
+    uiState: 'show'
+  });
 
   this.render(openTemplate);
 
-  assert.elementFound('.invitation-item-action-delete', 'destroy button is found');
+  assert.elementNotFound('.invitation-item-action-delete');
+  assert.elementFound('.invitation-item-action-rescind');
+  assert.elementNotFound('.invitation-item-action-send');
+});
 
+test('the row is in the show state, invitation is invited, and in previous round', function(assert) {
   this.setProperties({
-    'invitation.pending': true,
+    'invitation.state': 'invited',
     currentRound: false,
     uiState: 'show'
   });
 
-  assert.elementNotFound('.invitation-item-action-delete', 'destroy button not shown if not current round');
+  this.render(openTemplate);
 
+  assert.elementNotFound('.invitation-item-action-delete');
+  assert.elementNotFound('.invitation-item-action-rescind');
+  assert.elementNotFound('.invitation-item-action-send');
+});
+
+test('the row is in the show state, invitation is declined, and in current round', function(assert) {
   this.setProperties({
-    'invitation.pending': false,
+    'invitation.state': 'declined',
     currentRound: true,
     uiState: 'show'
   });
 
-  assert.elementNotFound('.invitation-item-action-delete', 'destroy button not shown if invitation is not pending');
+  this.render(openTemplate);
 
+  assert.elementNotFound('.invitation-item-action-delete');
+  assert.elementNotFound('.invitation-item-action-rescind');
+  assert.elementNotFound('.invitation-item-action-send');
+});
+
+test('grouped invitations are disabled when their primary has been invited or accepted', function(assert) {
+  this.set('invitation', make('invitation', {
+    email: 'jane@example.com',
+    invitee: { fullName: 'Jane McEdits' },
+    state: 'pending',
+    title: 'Awesome Paper!',
+    primary: make('invitation', {state: 'invited'})
+  }));
   this.setProperties({
-    'invitation.pending': true,
     currentRound: true,
-    uiState: 'closed'
+    uiState: 'show'
   });
 
-  assert.elementNotFound('.invitation-item-action-delete', 'destroy button not shown if the row is closed');
+  this.render(openTemplate);
+  assert.elementFound('.invitation-item--disabled');
+  this.set('invitation.primary.state', 'accepted');
+  assert.elementFound('.invitation-item--disabled');
+
+  this.set('invitation.primary.state', 'pending');
+  assert.elementNotFound('.invitation-item--disabled');
+  this.set('invitation.primary.state', 'declined');
+  assert.elementNotFound('.invitation-item--disabled');
 });
+
+function testAlternateState(state, shouldBeDisabled) {
+  test(`grouped invitations are ${shouldBeDisabled ? 'disabled' : 'enabled'} when any alternate in the group is ${state}`, function(assert) {
+    let alternate = make('invitation', {state: state});
+
+    this.set('invitation', make('invitation', {
+      email: 'jane@example.com',
+      invitee: { fullName: 'Jane McEdits' },
+      state: 'pending',
+      title: 'Awesome Paper!',
+      primary: make('invitation', {state: 'pending', alternates: [alternate]})
+    }));
+    this.setProperties({
+      currentRound: true,
+      uiState: 'show'
+    });
+
+    this.render(openTemplate);
+
+    if (shouldBeDisabled) {
+      assert.elementFound('.invitation-item--disabled');
+    } else {
+      assert.elementNotFound('.invitation-item--disabled');
+    }
+  });
+}
+
+testAlternateState('invited', true);
+testAlternateState('accepted', true);
+testAlternateState('pending', false);
+testAlternateState('declined', false);
 
 test('displays decline feedback when declined', function(assert){
   this.setProperties({
-    'invitation.declined': true,
+    'invitation.state': 'declined',
     'invitation.declineReason': 'No current availability',
     'invitation.reviewerSuggestions': 'Jane McReviewer'
   });