+++ spec/factories/invitation_queue_factory.rb
@@ -1,0 +1,4 @@
+FactoryGirl.define do
+  factory :invitation_queue do
+  end
+end